리팩토링: 파일 추출
